+++ product/infrastructure/fake/productservice.go
@@ -32,7 +32,7 @@ type ProductService struct {
 // Inject dependencies
 func (ps *ProductService) Inject(
 	c *struct {
-		CurrencyCode string `inject:"config:commerce.fakeservice.currency,optional"`
+		CurrencyCode string `inject:"config:commerce.product.fakeservice.currency,optional"`
 	},
 ) *ProductService {
 	if c != nil {